feat: add web interface for video-to-flashcards
Add a web interface to run `video-to-flashcards-cached` on a short
YouTube video (<15m) and produce a summary of the content in the form
of flashcards. The interface uses Gradio and allows users to enter the
URL of a video, and get the output text and a file, as well as status
messages and error logs. The interface is launched by running
`video-to-flashcards-webui.py`.

Note: this isn't quite working yet, need to finish debugging it.