feat(webchat): Add 'media' file type to prevent editing non-text files
A new file type, "media", is introduced for files with extensions that are
determined to be non-editable (i.e., not having a 'text/*' MIME type).
This prevents attempting to load binary files like PDFs or images into
the text editor.

To achieve this, `get_file_type` is now `async` as it must check the
file's MIME type. This change necessitated propagating `async/await`
throughout its call chain.

The `check_ok_to_edit` function was renamed to `check_cannot_edit`
and now returns an error string instead of throwing.